FfaFeaturePkg: Updates for 202511 (#99)
## Description

Updated to use 202511 branches of mu_basecore.
Applied changes to DSC to correspond with moved library locations and
changed file names.
Updated override hash for TimerLibEx.
Modify files to use FdtLib from MdePkg instead of libfdt from
EmbeddedPkg (deprecated in 202511 branches)

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested
Local CI using clangpdb under windows

## Integration Instructions
No integration necessary.

+++ FfaFeaturePkg/Library/ArmArchTimerLibEx/ArmArchTimerLibEx.c
@@ -56,6 +56,11 @@ MicroSecondDelay (
   UINT64  TimerTicks64;
   UINT64  SystemCounterVal;
 
+  // MU_CHANGE - Start - Cannot depend on GenericTimer Calls
+  //  UINT64  PreviousSystemCounterVal;
+  //  UINT64  DeltaCounterVal;
+  // MU_CHANGE - End - Cannot depend on GenericTimer Calls
+
   // Calculate counter ticks that represent requested delay:
   //  = MicroSeconds x TICKS_PER_MICRO_SEC
   //  = MicroSeconds x Frequency.10^-6
@@ -67,10 +72,24 @@ MicroSecondDelay (
                    1000000U
                    );
 
+  // MU_CHANGE - Start - Cannot depend on GenericTimer Calls
+  //  // Read System Counter value
+  //  PreviousSystemCounterVal = ArmGenericTimerGetSystemCount ();
+  //
+  //  // Wait until delay count expires.
+  //  while (TimerTicks64 > 0) {
+  //    SystemCounterVal = ArmGenericTimerGetSystemCount ();
+  //    // Get how much we advanced this tick. Wrap around still has delta correct
+  //    DeltaCounterVal = (SystemCounterVal - PreviousSystemCounterVal)
+  //                      & (MAX_UINT64 >> 8); // Account for a lesser (minimum) size
+  //    // Never wrap back around below zero by choosing the min and thus stop at 0
+  //    TimerTicks64            -= MIN (TimerTicks64, DeltaCounterVal);
+  //    PreviousSystemCounterVal = SystemCounterVal;
   // Wait until delay count expires.
   SystemCounterVal = 0;
   while (SystemCounterVal < TimerTicks64) {
     SystemCounterVal++;
+    // MU_CHANGE - End - Cannot depend on GenericTimer Calls
   }
 
   return MicroSeconds;
@@ -191,7 +210,7 @@ GetTimeInNanoSecond (
   UINT64  Remainder;
   UINT64  TimerFreq;
 
-  TimerFreq = (UINT32)GetPlatformTimerFreq ();  // MU_CHANGE - ARM64 VS change
+  TimerFreq = GetPlatformTimerFreq ();
   //
   //          Ticks
   // Time = --------- x 1,000,000,000
